refactor(polarsdf): move private helpers to functions_helpers
Keeps sparkleframe.polarsdf.functions focused on the public Spark-like API;
underscore helpers and shared format/JSON/timestamp utilities live alongside
functions_utils.py. Lazy-imports col() inside one-arg timestamp helper to avoid
import cycles.
